! [Fixed] CyclesOfGraph.h - fixed guard include definition naming (set equal to the header name), dependent includes and namespace using directive moved to CyclesOfGraph.cpp for better encapsulation, NodesOfGraph header declaration replaced by forward, NodesOfGraph declaration moved to CyclesOfGraph.cpp.
